requirements.txt
praw==7.7.1
facebook-sdk==3.1.0
tweepy==4.14.0
python-dotenv==1.0.0
openai==1.3.0
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2 